fix: address PR #717 CI failure
